Move more step_* code to dedicated modules, standardize as methods
