+++ crates/kitsune_p2p/kitsune_p2p/src/gossip/simple_bloom/step_4_com_loop_inner.rs
@@ -1,6 +1,96 @@
 use super::*;
 use kitsune_p2p_types::codec::*;
 
+impl SimpleBloomMod {
+    pub(crate) async fn step_4_com_loop_inner(&self) -> KitsuneResult<()> {
+        let loop_start = std::time::Instant::now();
+
+        loop {
+            let (tuning_params, space, ep_hnd) = (
+                self.tuning_params.clone(),
+                self.space.clone(),
+                self.ep_hnd.clone(),
+            );
+            let (mut maybe_outgoing, mut maybe_incoming) =
+                self.inner.share_mut(|i, _| {
+                    let maybe_outgoing = if !i.outgoing.is_empty()
+                        && i.last_outgoing.elapsed().as_millis() as u64 > self.send_interval_ms
+                    {
+                        let (cert, how, gossip) = i.outgoing.remove(0);
+
+                        // set this to a time in the future
+                        // so we don't accidentally double up if sending
+                        // is slow... we'll set this more reasonably
+                        // when we get a success or failure below.
+                        i.last_outgoing = std::time::Instant::now()
+                            .checked_add(std::time::Duration::from_millis(
+                                self.tuning_params.tx2_implicit_timeout_ms as u64,
+                            ))
+                            .expect("Congratulations on running holochain near the heat death of the universe :)");
+
+                        Some((cert, how, gossip))
+                    } else {
+                        None
+                    };
+                    let maybe_incoming = if !i.incoming.is_empty() {
+                        Some(i.incoming.remove(0))
+                    } else {
+                        None
+                    };
+                    Ok((
+                        maybe_outgoing,
+                        maybe_incoming,
+                    ))
+                })?;
+
+            let will_break = (maybe_outgoing.is_none() && maybe_incoming.is_none())
+                || loop_start.elapsed().as_millis() as u32
+                    > tuning_params.gossip_loop_iteration_delay_ms;
+
+            if let Some(outgoing) = maybe_outgoing.take() {
+                let (endpoint, how, gossip) = outgoing;
+                let agents = endpoint.agents().clone();
+                if let Err(e) = step_4_com_loop_inner_outgoing(
+                    &self.inner,
+                    tuning_params.clone(),
+                    space.clone(),
+                    ep_hnd,
+                    endpoint.cert().clone(),
+                    how,
+                    gossip,
+                )
+                .await
+                {
+                    tracing::warn!("failed to send outgoing: {:?} {:?}", endpoint, e);
+                    self.inner.share_mut(move |i, _| {
+                        i.last_outgoing = std::time::Instant::now();
+                        i.record_pending_metric(agents, true);
+                        Ok(())
+                    })?;
+                } else {
+                    self.inner.share_mut(move |i, _| {
+                        i.last_outgoing = std::time::Instant::now();
+                        i.record_pending_metric(agents, false);
+                        Ok(())
+                    })?;
+                }
+            }
+
+            if let Some(incoming) = maybe_incoming.take() {
+                let (con, gossip) = incoming;
+                if let Err(e) = step_4_com_loop_inner_incoming(&self, con, gossip).await {
+                    tracing::warn!("failed to process incoming: {:?}", e);
+                }
+            }
+
+            if will_break {
+                break;
+            }
+        }
+        Ok(())
+    }
+}
+
 pub(crate) async fn step_4_com_loop_inner_outgoing(
     inner: &Share<SimpleBloomModInner>,
     tuning_params: KitsuneP2pTuningParams,
